implemented git and vscode

+++ script.py
@@ -1,11 +1,27 @@
 import subprocess, os, shutil, sys
 from settings import *
 
+STEPS = """
+- get input
+- create folder (project name)
+- create the following empty files
+- create backend project folder (django-admin startproject backend)
+- create empty files in backend folder (.gitignore, packages.txt)
+- create app "base"
+- update backend's .gitignore
+- setup basic urls
+- create frontend project folder (npx create-vite@latest frontend --template react)
+- install tailwind
+- setup tailwind
+- remove bloatware from frontend like CSS files
+- git init
+- open vscodes
+- delete __pycache__ created from this script
+"""
+
 def replace_file_contents(filepath, lines):
     try:
         with open(filepath, "w") as file:
-            file.write("")
-        with open(filepath, "a") as file:
             for line in lines:
                 file.write(line + "\n")
         return True
@@ -25,11 +41,18 @@ def run_cmd(cwd, cmd, shell):
         return False
     return True
 
+def run_cmds(cwd, cmds, shell):
+    for cmd in cmds:
+        if not run_cmd(cwd, cmd, shell):
+            return False
+    return True
+
 def create_fullstack_project(PROJECT_NAME):
     global TAKE_INPUT, OVERRIDE_EXISTING_PROJECT_FILENAME, DJANGO_PROJECT_NAME, DJANGO_APP_NAME, REACT_PROJECT_NAME, PROJECT_URLS_PY_LINES, APP_URLS_PY_LINES, APP_VIEWS_PY_LINES, DJANGO_PACKAGES_LINES, VITE_CONFIG_JS_LINES, INDEX_HTML_LINES, APP_JSX_LINES, MAIN_JSX_LINES
 
     # get input
     if TAKE_INPUT:
+        GIT_COMMIT_NAME = input("git commit name: ")
         DJANGO_PROJECT_NAME = input("django project name: ")
         DJANGO_APP_NAME = input("django app name: ")
         REACT_PROJECT_NAME = input("react project name: ")
@@ -170,9 +193,36 @@ def create_fullstack_project(PROJECT_NAME):
             return
     print(f"frontend bloatware removed")
 
+    # git init
+    if INITIALIZE_GIT:
+        GIT_COMMIT_NAME = "\"" + GIT_COMMIT_NAME + "\""
+        if not run_cmds(
+            cwd=PROJECT_NAME, shell=False,
+            cmd=[
+                ["git", "init"],
+                ["git", "add", "."],
+                ["git", "commit", "-m", GIT_COMMIT_NAME],
+            ]
+        ): return
+        print(f"git initialized")
+
+    # open vscodes
+    backend_dir = os.path.join(PROJECT_NAME, "apps", DJANGO_PROJECT_NAME)
+    frontend_dir = os.path.join(PROJECT_NAME, "apps", REACT_PROJECT_NAME)
+    if OPEN_VSCODES:
+        try:
+            if not run_cmd(
+                cwd=backend_dir, shell=False,
+                cmd=["code", "."],
+            ) or not run_cmd(
+                cwd=frontend_dir, shell=False,
+                cmd=["code", "."],
+            ): return
+        except Exception as e:
+            print(f"couldnt open vscode: {e}")
+
     # delete __pycache__ created from this script
-    if os.path.isdir("__pycache__"):
-        shutil.rmtree("__pycache__")
+    if os.path.isdir("__pycache__"): shutil.rmtree("__pycache__")
 
 if __name__ == "__main__":
     PROJECT_NAME = sys.argv[1] if len(sys.argv) > 1 else input("project name: ")
